requirements.txt
alive_progress==3.1.4
numpy==1.25.2
opencv_python==4.8.0.76
pandas==2.0.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
matplotlib==3.8.0
